pyproject.toml
[tool.poetry]
name = "epubtranslator"
version = "1.0.0"
description = ""
authors = ["ShenSheiBot <shensheibot@proton.me>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.3"
EbookLib = "^0.18"
fastapi_poe = "^0.0.32"
google_api_python_client = "^2.99.0"
google_auth_oauthlib = "^1.1.0"
hanziconv = "^0.3.2"
loguru = "^0.7.2"
openai = "1.12.0"
protobuf = "^4.25.2"
pyperclip = "^1.8.2"
PyYAML = "^6.0.1"
selenium = "^4.17.2"
tqdm = "^4.66.1"
undetected_chromedriver = "^3.5.3"
google-generativeai = "^0.3.2"
pyzipper = "^0.3.6"
jupyterlab = "^4.0.10"
transformers = "^4.36.2"
ipywidgets = "^8.1.1"
sentencepiece = "^0.1.99"
google-cloud-vision = "^3.5.0"
matplotlib = "^3.8.2"
scikit-learn = "^1.3.2"
opencv-python = "^4.9.0.80"
markdown = "^3.5.2"
shutils = "^0.1.0"
accelerate = "^0.26.1"
bitsandbytes = "^0.42.0"
torch = {version = "^2.2.0+cu121", source = "pytorch-gpu-src"}
torchvision = {version = "^0.17.0+cu121", source = "pytorch-gpu-src"}
torchaudio = {version = "^2.2.0+cu121", source = "pytorch-gpu-src"}
ja-sentence-segmenter = "^0.0.2"
starlette-context = "^0.3.6"
pydantic-settings = "^2.2.1"

[[tool.poetry.source]]
name = "pytorch-gpu-src"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
